self-healing-neurosym-trajectory-repair/requirements.txt
torch>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
httpx>=0.24.0
structlog>=23.0.0
docker>=6.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
